requirements.txt
numpy
matplotlib
librosa
python-dotenv
Pillow